backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
numpy
python-dotenv
joblib
requests